Refactor graphql filters
