+++ saleor/graphql/checkout/filters.py
@@ -11,13 +11,14 @@ from ..channel.filters import get_currency_from_filter_data
 from ..channel.types import Channel
 from ..core.doc_category import DOC_CATEGORY_CHECKOUT
 from ..core.filters import (
+    FilterInputObjectType,
     GlobalIDMultipleChoiceFilter,
     ListObjectTypeFilter,
     MetadataFilter,
     MetadataFilterBase,
     ObjectTypeFilter,
 )
-from ..core.types import DateRangeInput, FilterInputObjectType
+from ..core.types import DateRangeInput
 from ..core.utils import from_global_id_or_error
 from ..discount.filters import DiscountedObjectWhere
 from ..utils import resolve_global_ids_to_primary_keys
