+++ saleor/graphql/core/types/__init__.py
@@ -75,14 +75,6 @@ from .common import (
     Weight,
 )
 from .event import SubscriptionObjectType
-from .filter_input import (
-    ChannelFilterInputObjectType,
-    DateFilterInput,
-    DateTimeFilterInput,
-    FilterInputObjectType,
-    IntFilterInput,
-    StringFilterInput,
-)
 from .model import ModelObjectType
 from .money import VAT, Money, MoneyRange, ReducedRate, TaxedMoney, TaxedMoneyRange
 from .sort_input import ChannelSortInputObjectType, SortInputObjectType
@@ -171,16 +163,10 @@ __all__ = [
     "WebhookError",
     "WebhookDryRunError",
     "WebhookTriggerError",
-    "FilterInputObjectType",
     "SortInputObjectType",
-    "ChannelFilterInputObjectType",
     "ChannelSortInputObjectType",
     "Upload",
     "TYPES_WITH_DOUBLE_ID_AVAILABLE",
-    "StringFilterInput",
-    "IntFilterInput",
-    "DateFilterInput",
-    "DateTimeFilterInput",
     "IconThumbnailField",
     "VoucherCodeBulkDeleteError",
 ]
