+++ saleor/graphql/webhook/filters.py
@@ -1,7 +1,6 @@
 import django_filters
 
-from ..core.filters import EnumFilter
-from ..core.types import FilterInputObjectType
+from ..core.filters import EnumFilter, FilterInputObjectType
 from .enums import EventDeliveryStatusEnum, WebhookEventTypeEnum
